requirements.txt
# Streamlit for the frontend
streamlit

# Backend libraries
fastapi
uvicorn

# OPEA components
python-multipart
# pydantic>=1.8.2
# requests>=2.26.0
opea-comps
httpx

# SQLite and ChromaDB for database management
sqlite-utils
aiosqlite
chromadb

# YouTube transcript downloader
youtube-transcript-api

# Automatic Speech Recognition (ASR)
openwhisper
openai-whisper

# Text-to-Speech (Coqui TTS)
coqui-tts
coqui-tts[ja]
TTS
torch
numpy
soundfile

# LLM interaction (requests for API calls)
requests

# Content validation and regex
regex

# Additional utilities
pydantic
python-dotenv
ffmpeg
pykakasi
pytest
pytest-asyncio

# Docker and NVIDIA support (if needed for GPU)
docker
nvidia-pyindex
# nvidia-tensorflow==23.9

# Optional: For development and linting
black
flake8

streamlit>=1.22.0
requests>=2.28.2
python-dotenv>=1.0.0
Pillow>=9.5.0
manga-ocr==0.1.8
transformers==4.30.2
torch==2.0.1